HMW-131 Fix git diff part 1 of 2.
